+++ lib/playx_widget.dart
@@ -2,4 +2,5 @@ library playx_widget;
 
 export 'package:cached_network_image/cached_network_image.dart';
 export 'package:flutter_svg/svg.dart';
+export 'package:lottie/lottie.dart';
 export 'package:package_info_plus/package_info_plus.dart';