+++ CHANGELOG.md
@@ -1,2 +1,9 @@
+## 0.0.2
+- Add support for Dart 3.0.0 
+- Upgrade Packages.
+- Add [AutoAnimatedList] : Widget that automatically animates list of items when inserting or deleting any item.
+- Add [OptimizedCard] : Widget which provides better shadow effect for the card to look more like native card.
+- Add lottie package. 
+
 ## 0.0.1
 - initial release